split previous nxdata module into 4 submodules _utils, parse, validate, write
