+++ silx/io/nxdata/parse.py
@@ -22,294 +22,22 @@
 # THE SOFTWARE.
 #
 # ###########################################################################*/
-"""This module provides a collection of functions to work with h5py-like
-groups following the NeXus *NXdata* specification.
-
-See http://download.nexusformat.org/sphinx/classes/base_classes/NXdata.html
-
-"""
 import logging
-import os
-import os.path
 import numpy
-from silx.io.utils import is_dataset, is_group, is_file
+from silx.io.utils import is_dataset, is_group
+from .validate import is_valid_nxdata,\
+    is_NXroot_with_default_NXdata,\
+    is_NXentry_with_default_NXdata
+from .validate import _INTERPDIM
+from ._utils import get_attr_as_unicode
 from silx.third_party import six
-from silx.utils.deprecation import deprecated
-
-try:
-    import h5py
-except ImportError:
-    h5py = None
 
 __authors__ = ["P. Knobel"]
 __license__ = "MIT"
-__date__ = "12/02/2018"
+__date__ = "17/04/2018"
+
 
 _logger = logging.getLogger(__name__)
-
-
-_INTERPDIM = {"scalar": 0,
-              "spectrum": 1,
-              "image": 2,
-              "rgba-image": 3,  # "hsla-image": 3, "cmyk-image": 3, # TODO
-              "vertex": 1}  # 3D scatter: 1D signal + 3 axes (x, y, z) of same legth
-"""Number of signal dimensions associated to each possible @interpretation
-attribute.
-"""
-
-
-def _nxdata_warning(msg, group_name=""):
-    """Log a warning message prefixed with
-    *"NXdata warning: "*
-
-    :param str msg: Warning message
-    :param str group_name: Name of NXdata group this warning relates to
-    """
-    warning_prefix = "NXdata warning"
-    if group_name:
-        warning_prefix += " (group %s): " % group_name
-    else:
-        warning_prefix += ": "
-    _logger.warning(warning_prefix + msg)
-
-
-@deprecated(since_version="0.8.0", replacement="get_attr_as_unicode")
-def get_attr_as_string(*args, **kwargs):
-    return get_attr_as_unicode(*args, **kwargs)
-
-
-def get_attr_as_unicode(item, attr_name, default=None):
-    """Return item.attrs[attr_name] as unicode or as a
-    list of unicode.
-
-    Numpy arrays of strings or bytes returned by h5py are converted to
-    lists of unicode.
-
-    :param item: Group or dataset
-    :param attr_name: Attribute name
-    :param default: Value to be returned if attribute is not found.
-    :return: item.attrs[attr_name]
-    """
-    attr = item.attrs.get(attr_name, default)
-
-    if isinstance(attr, six.binary_type):
-        # byte-string
-        return attr.decode("utf-8")
-    elif isinstance(attr, numpy.ndarray) and not attr.shape:
-        if isinstance(attr[()], six.binary_type):
-            # byte string as ndarray scalar
-            return attr[()].decode("utf-8")
-        else:
-            # other scalar, possibly unicode
-            return attr[()]
-    elif isinstance(attr, numpy.ndarray) and len(attr.shape):
-        if hasattr(attr[0], "decode"):
-            # array of byte-strings
-            return [element.decode("utf-8") for element in attr]
-        else:
-            # other array, most likely unicode objects
-            return [element for element in attr]
-    else:
-        return attr
-
-
-def is_valid_nxdata(group):   # noqa
-    """Check if a h5py group is a **valid** NX_data group.
-
-    If the group does not have attribute *@NX_class=NXdata*, this function
-    simply returns *False*.
-
-    Else, warning messages are logged to troubleshoot malformed NXdata groups
-    prior to returning *False*.
-
-    :param group: h5py-like group
-    :return: True if this NXdata group is valid.
-    :raise TypeError: if group is not a h5py group, a spech5 group,
-        or a fabioh5 group
-    """
-    if not is_group(group):
-        raise TypeError("group must be a h5py-like group")
-    if get_attr_as_unicode(group, "NX_class") != "NXdata":
-        return False
-    if "signal" not in group.attrs:
-        _logger.info("NXdata group %s does not define a signal attr. "
-                     "Testing legacy specification.", group.name)
-        signal_name = None
-        for key in group:
-            if "signal" in group[key].attrs:
-                signal_name = key
-                signal_attr = group[key].attrs["signal"]
-                if signal_attr in [1, b"1", u"1"]:
-                    # This is the main (default) signal
-                    break
-        if signal_name is None:
-            _nxdata_warning("No @signal attribute on the NXdata group, "
-                            "and no dataset with a @signal=1 attr found",
-                            group.name)
-            return False
-    else:
-        signal_name = get_attr_as_unicode(group, "signal")
-
-    if signal_name not in group or not is_dataset(group[signal_name]):
-        _nxdata_warning(
-            "Cannot find signal dataset '%s'" % signal_name,
-            group.name)
-        return False
-
-    auxiliary_signals_names = get_attr_as_unicode(group, "auxiliary_signals",
-                                                  default=[])
-    if isinstance(auxiliary_signals_names, (six.text_type, six.binary_type)):
-        auxiliary_signals_names = [auxiliary_signals_names]
-    for asn in auxiliary_signals_names:
-        if asn not in group or not is_dataset(group[asn]):
-            _nxdata_warning(
-                "Cannot find auxiliary signal dataset '%s'" % asn,
-                group.name)
-            return False
-        if group[signal_name].shape != group[asn].shape:
-            _nxdata_warning("Auxiliary signal dataset '%s' does not" % asn +
-                            " have the same shape as the main signal.",
-                            group.name)
-            return False
-
-    ndim = len(group[signal_name].shape)
-
-    if "axes" in group.attrs:
-        axes_names = get_attr_as_unicode(group, "axes")
-        if isinstance(axes_names, (six.text_type, six.binary_type)):
-            axes_names = [axes_names]
-
-        if 1 < ndim < len(axes_names):
-            # ndim = 1 with several axes could be a scatter
-            _nxdata_warning(
-                "More @axes defined than there are " +
-                "signal dimensions: " +
-                "%d axes, %d dimensions." % (len(axes_names), ndim),
-                group.name)
-            return False
-
-        # case of less axes than dimensions: number of axes must match
-        # dimensionality defined by @interpretation
-        if ndim > len(axes_names):
-            interpretation = get_attr_as_unicode(group[signal_name], "interpretation")
-            if interpretation is None:
-                interpretation = get_attr_as_unicode(group, "interpretation")
-            if interpretation is None:
-                _nxdata_warning("No @interpretation and not enough" +
-                                " @axes defined.", group.name)
-                return False
-
-            if interpretation not in _INTERPDIM:
-                _nxdata_warning("Unrecognized @interpretation=" + interpretation +
-                                " for data with wrong number of defined @axes.",
-                                group.name)
-                return False
-            if interpretation == "rgba-image":
-                if ndim != 3 or group[signal_name].shape[-1] not in [3, 4]:
-                    _nxdata_warning(
-                        "Inconsistent RGBA Image. Expected 3 dimensions with " +
-                        "last one of length 3 or 4. Got ndim=%d " % ndim +
-                        "with last dimension of length %d." % group[signal_name].shape[-1],
-                        group.name)
-                    return False
-                if len(axes_names) != 2:
-                    _nxdata_warning(
-                        "Inconsistent number of axes for RGBA Image. Expected "
-                        "3, but got %d." % ndim, group.name)
-                    return False
-
-            elif len(axes_names) != _INTERPDIM[interpretation]:
-                _nxdata_warning(
-                    "%d-D signal with @interpretation=%s " % (ndim, interpretation) +
-                    "must define %d or %d axes." % (ndim, _INTERPDIM[interpretation]),
-                    group.name)
-                return False
-
-        # Test consistency of @uncertainties
-        uncertainties_names = get_attr_as_unicode(group, "uncertainties")
-        if uncertainties_names is None:
-            uncertainties_names = get_attr_as_unicode(group[signal_name], "uncertainties")
-        if isinstance(uncertainties_names, str):
-            uncertainties_names = [uncertainties_names]
-        if uncertainties_names is not None:
-            if len(uncertainties_names) != len(axes_names):
-                _nxdata_warning("@uncertainties does not define the same " +
-                                "number of fields than @axes", group.name)
-                return False
-
-        # Test individual axes
-        is_scatter = True   # true if all axes have the same size as the signal
-        signal_size = 1
-        for dim in group[signal_name].shape:
-            signal_size *= dim
-        polynomial_axes_names = []
-        for i, axis_name in enumerate(axes_names):
-
-            if axis_name == ".":
-                continue
-            if axis_name not in group or not is_dataset(group[axis_name]):
-                _nxdata_warning("Could not find axis dataset '%s'" % axis_name,
-                                group.name)
-                return False
-
-            axis_size = 1
-            for dim in group[axis_name].shape:
-                axis_size *= dim
-
-            if len(group[axis_name].shape) != 1:
-                # too me, it makes only sense to have a n-D axis if it's total
-                # size is exactly the signal's size (weird n-d scatter)
-                if axis_size != signal_size:
-                    _nxdata_warning("Axis %s is not a 1D dataset" % axis_name +
-                                    " and its shape does not match the signal's shape",
-                                    group.name)
-                    return False
-                axis_len = axis_size
-            else:
-                # for a  1-d axis,
-                fg_idx = group[axis_name].attrs.get("first_good", 0)
-                lg_idx = group[axis_name].attrs.get("last_good", len(group[axis_name]) - 1)
-                axis_len = lg_idx + 1 - fg_idx
-
-            if axis_len != signal_size:
-                if axis_len not in group[signal_name].shape + (1, 2):
-                    _nxdata_warning(
-                        "Axis %s number of elements does not " % axis_name +
-                        "correspond to the length of any signal dimension,"
-                        " it does not appear to be a constant or a linear calibration," +
-                        " and this does not seem to be a scatter plot.", group.name)
-                    return False
-                elif axis_len in (1, 2):
-                    polynomial_axes_names.append(axis_name)
-                is_scatter = False
-            else:
-                if not is_scatter:
-                    _nxdata_warning(
-                        "Axis %s number of elements is equal " % axis_name +
-                        "to the length of the signal, but this does not seem" +
-                        " to be a scatter (other axes have different sizes)",
-                        group.name)
-                    return False
-
-            # Test individual uncertainties
-            errors_name = axis_name + "_errors"
-            if errors_name not in group and uncertainties_names is not None:
-                errors_name = uncertainties_names[i]
-                if errors_name in group and axis_name not in polynomial_axes_names:
-                    if group[errors_name].shape != group[axis_name].shape:
-                        _nxdata_warning(
-                            "Errors '%s' does not have the same " % errors_name +
-                            "dimensions as axis '%s'." % axis_name, group.name)
-                        return False
-
-    # test dimensions of errors associated with signal
-    if "errors" in group and is_dataset(group["errors"]):
-        if group["errors"].shape != group[signal_name].shape:
-            _nxdata_warning("Dataset containing standard deviations must " +
-                            "have the same dimensions as the signal.",
-                            group.name)
-            return False
-    return True
 
 
 class NXdata(object):
@@ -780,7 +508,7 @@ class NXdata(object):
         if self.signal_is_0d or self.signal_is_1d:
             return False
         if not self.signal_is_2d and \
-                        self.interpretation not in ["image", "rgba-image"]:
+                self.interpretation not in ["image", "rgba-image"]:
             return False
         if self.signal_is_3d and self.interpretation == "rgba-image":
             if self.signal.shape[-1] not in [3, 4]:
@@ -813,48 +541,6 @@ class NXdata(object):
         return True
 
 
-def is_NXentry_with_default_NXdata(group):
-    """Return True if group is a valid NXentry defining a valid default
-    NXdata."""
-    if not is_group(group):
-        return False
-
-    if get_attr_as_unicode(group, "NX_class") != "NXentry":
-        return False
-
-    default_nxdata_name = group.attrs.get("default")
-    if default_nxdata_name is None or default_nxdata_name not in group:
-        return False
-
-    default_nxdata_group = group.get(default_nxdata_name)
-
-    if not is_group(default_nxdata_group):
-        return False
-
-    return is_valid_nxdata(default_nxdata_group)
-
-
-def is_NXroot_with_default_NXdata(group):
-    """Return True if group is a valid NXroot defining a default NXentry
-    defining a valid default NXdata."""
-    if not is_group(group):
-        return False
-
-    # A NXroot is supposed to be at the root of a data file, and @NX_class
-    # is therefore optional. We accept groups that are not located at the root
-    # if they have @NX_class=NXroot (use case: several nexus files archived
-    # in a single HDF5 file)
-    if get_attr_as_unicode(group, "NX_class") != "NXroot" and not is_file(group):
-        return False
-
-    default_nxentry_name = group.attrs.get("default")
-    if default_nxentry_name is None or default_nxentry_name not in group:
-        return False
-
-    default_nxentry_group = group.get(default_nxentry_name)
-    return is_NXentry_with_default_NXdata(default_nxentry_group)
-
-
 def get_default(group):
     """Return a :class:`NXdata` object corresponding to the default NXdata group
     in the group specified as parameter.
@@ -884,168 +570,3 @@ def get_default(group):
         return None
 
     return NXdata(default_data)
-
-
-def _str_to_utf8(text):
-    return numpy.array(text, dtype=h5py.special_dtype(vlen=six.text_type))
-
-
-def save_NXdata(filename, signal, axes=None,
-                signal_name="data", axes_names=None,
-                signal_long_name=None, axes_long_names=None,
-                signal_errors=None, axes_errors=None,
-                title=None, interpretation=None,
-                nxentry_name="entry", nxdata_name=None):
-    """Write data to an NXdata group.
-
-    .. note::
-
-        No consistency checks are made regarding the dimensionality of the
-        signal and number of axes. The user is responsible for providing
-        meaningful data, that can be interpreted by visualization software.
-
-    :param str filename: Path to output file. If the file does not
-        exists, it is created.
-    :param numpy.ndarray signal: Signal array.
-    :param List[numpy.ndarray] axes: List of axes arrays.
-    :param str signal_name: Name of signal dataset, in output file
-    :param List[str] axes_names: List of dataset names for axes, in
-        output file
-    :param str signal_long_name: *@long_name* attribute for signal, or None.
-    :param  axes_long_names: None, or list of long names
-        for axes
-    :type axes_long_names: List[str, None]
-    :param numpy.ndarray signal_errors: Array of errors associated with the
-        signal
-    :param axes_errors: List of arrays of errors
-        associated with each axis
-    :type axes_errors: List[numpy.ndarray, None]
-    :param str title: Graph title (saved as a "title" dataset) or None.
-    :param str interpretation: *@interpretation* attribute ("spectrum",
-        "image", "rgba-image" or None). This is only needed in cases of
-        ambiguous dimensionality, e.g. a 3D array which represents a RGBA
-        image rather than a stack.
-    :param str nxentry_name: Name of group in which the NXdata group
-        is created. By default, "/entry" is used.
-
-        .. note::
-
-            The Nexus format specification requires for NXdata groups
-            be part of a NXentry group.
-            The specified group should have attribute *@NX_class=NXentry*, in
-            order for the created file to be nexus compliant.
-    :param str nxdata_name: Name of NXdata group. If omitted (None), the
-        function creates a new group using the first available name ("data0",
-        or "data1"...).
-        Overwriting an existing group (or dataset) is not supported, you must
-        delete it yourself prior to calling this function if this is what you
-        want.
-    :return: True if save was successful, else False.
-    """
-    if h5py is None:
-        raise ImportError("h5py could not be imported, but is required by "
-                          "save_NXdata function")
-
-    if axes_names is not None:
-        assert axes is not None, "Axes names defined, but missing axes arrays"
-        assert len(axes) == len(axes_names), \
-            "Mismatch between number of axes and axes_names"
-
-    if axes is not None and axes_names is None:
-        axes_names = []
-        for i, axis in enumerate(axes):
-            axes_names.append("dim%d" % i if axis is not None else ".")
-    if axes is None:
-        axes = []
-
-    # Open file in
-    if os.path.exists(filename):
-        errmsg = "Cannot write/append to existing path %s"
-        if not os.path.isfile(filename):
-            errmsg += " (not a file)"
-            _logger.error(errmsg, filename)
-            return False
-        if not os.access(filename, os.W_OK):
-            errmsg += " (no permission to write)"
-            _logger.error(errmsg, filename)
-            return False
-        mode = "r+"
-    else:
-        mode = "w-"
-
-    with h5py.File(filename, mode=mode) as h5f:
-        # get or create entry
-        if nxentry_name is not None:
-            entry = h5f.require_group(nxentry_name)
-            if "default" not in h5f.attrs:
-                # set this entry as default
-                h5f.attrs["default"] = _str_to_utf8(nxentry_name)
-            if "NX_class" not in entry.attrs:
-                entry.attrs["NX_class"] = u"NXentry"
-        else:
-            # write NXdata into the root of the file (invalid nexus!)
-            entry = h5f
-
-        # Create NXdata group
-        if nxdata_name is not None:
-            if nxdata_name in entry:
-                _logger.error("Cannot assign an NXdata group to an existing"
-                              " group or dataset")
-                return False
-        else:
-            # no name specified, take one that is available
-            nxdata_name = "data0"
-            i = 1
-            while nxdata_name in entry:
-                _logger.info("%s item already exists in NXentry group," +
-                             " trying %s", nxdata_name, "data%d" % i)
-                nxdata_name = "data%d" % i
-                i += 1
-
-        data_group = entry.create_group(nxdata_name)
-        data_group.attrs["NX_class"] = u"NXdata"
-        data_group.attrs["signal"] = _str_to_utf8(signal_name)
-        if axes:
-            data_group.attrs["axes"] = _str_to_utf8(axes_names)
-        if title:
-            # not in NXdata spec, but implemented by nexpy
-            data_group["title"] = title
-            # better way imho
-            data_group.attrs["title"] = _str_to_utf8(title)
-
-        signal_dataset = data_group.create_dataset(signal_name,
-                                                   data=signal)
-        if signal_long_name:
-            signal_dataset.attrs["long_name"] = _str_to_utf8(signal_long_name)
-        if interpretation:
-            signal_dataset.attrs["interpretation"] = _str_to_utf8(interpretation)
-
-        for i, axis_array in enumerate(axes):
-            if axis_array is None:
-                assert axes_names[i] in [".", None], \
-                    "Axis name defined for dim %d but no axis array" % i
-                continue
-            axis_dataset = data_group.create_dataset(axes_names[i],
-                                                     data=axis_array)
-            if axes_long_names is not None:
-                axis_dataset.attrs["long_name"] = _str_to_utf8(axes_long_names[i])
-
-        if signal_errors is not None:
-            data_group.create_dataset("errors",
-                                      data=signal_errors)
-
-        if axes_errors is not None:
-            assert isinstance(axes_errors, (list, tuple)), \
-                "axes_errors must be a list or a tuple of ndarray or None"
-            assert len(axes_errors) == len(axes_names), \
-                "Mismatch between number of axes_errors and axes_names"
-            for i, axis_errors in enumerate(axes_errors):
-                if axis_errors is not None:
-                    dsname = axes_names[i] + "_errors"
-                    data_group.create_dataset(dsname,
-                                              data=axis_errors)
-        if "default" not in entry.attrs:
-            # set this NXdata as default
-            entry.attrs["default"] = nxdata_name
-
-    return True
